pyproject.toml
[project]
name = "irl-tools"
version = "0.1.0"
description = ""
authors = [
    {name = "nbebeshina",email = "nadia.clairet@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "gql (>=4.0.0,<5.0.0)",
    "poetry-core (>=2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
